mfd: intel_soc_pmic: set reference backlight PWM period
The reference PWM period should be set in the pwm_lookup entry instead of
assigning it in the PWM user driver (in our case the i915 DRM driver).
Specify the reference period here to get rid of the CRC_PMIC_PWM_PERIOD_NS
definition and rely on pwm_get_args() to retrieve the reference config.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>