drm: i915: switch to the atomic PWM API
Use the atomic API wherever appropriate and get rid of pwm_apply_args()
call (the reference period and polarity are now explicitly set when
calling pwm_apply_state()).

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>
CC: Jani Nikula <jani.nikula@intel.com>
CC: Shobhit Kumar <shobhit.kumar@intel.com>
CC: Boris Brezillon <boris.brezillon@free-electrons.com>

---
Hi,

Sorry guys, but I had to drop the Reviewed/Acked-by tags because this
version is completely different from the previous one. I now make full
usage of the atomic API to retrieve the initial PWM state and apply new
configs atomically.

Still unsure why we have to wait 2ms between the duty_cycle = 0
and enabled = false config. If that's a constraint of the PWM device
then it should probably be done in the PWM driver, but if it's related
to the panel itself, then it should stay in the DRM driver.

Thanks,

Boris